adapted dtype and device according to source code.

+++ tests/mesh/test_elements.py
@@ -12,57 +12,57 @@ class TestElements(unittest.TestCase):
             Elements(triangles=[1, 2, 3])
 
     def tests_if_type_error_is_raised_when_triangles_is_set_to_list(self):
-        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         with self.assertRaises(TypeError):
             elements.triangles = [1, 2, 3]
 
     def tests_if_value_error_is_raised_when_triangles_is_initialized_to_1d_tensor(self):
         with self.assertRaises(ValueError):
-            tri = torch.arange(12).to(dtype=torch.int64)
+            tri = torch.arange(12).to(dtype=torch.int32)
             Elements(triangles=tri)
 
     def tests_if_value_error_is_raised_when_triangles_is_set_to_1d_tensor(self):
-        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         with self.assertRaises(ValueError):
-            elements.triangles = torch.arange(12).to(dtype=torch.int64)
+            elements.triangles = torch.arange(12).to(dtype=torch.int32)
 
     def tests_if_value_error_is_raised_when_triangles_is_initialized_to_3d_tensor(self):
         with self.assertRaises(ValueError):
-            tri = torch.arange(12).reshape(2, 2, 3).to(dtype=torch.int64)
+            tri = torch.arange(12).reshape(2, 2, 3).to(dtype=torch.int32)
             Elements(triangles=tri)
 
     def tests_if_value_error_is_raised_when_triangles_is_set_to_3d_tensor(self):
-        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         with self.assertRaises(ValueError):
-            elements.triangles = tri.reshape(2, 2, 3).to(dtype=torch.int64)
+            elements.triangles = tri.reshape(2, 2, 3).to(dtype=torch.int32)
 
     def tests_if_value_error_is_raised_when_triangles_last_dimension_is_initialized_to_2(self):
         with self.assertRaises(ValueError):
-            tri = torch.arange(12).reshape(6, 2).to(dtype=torch.int64)
+            tri = torch.arange(12).reshape(6, 2).to(dtype=torch.int32)
             Elements(triangles=tri)
 
     def tests_if_value_error_is_raised_when_triangles_last_dimension_is_set_to_2(self):
-        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         with self.assertRaises(ValueError):
-            elements.triangles = torch.arange(12).reshape(6, 2).to(dtype=torch.int64)
+            elements.triangles = torch.arange(12).reshape(6, 2).to(dtype=torch.int32)
 
     def tests_if_value_error_is_raised_when_tetrahedras_last_dimension_is_initialized_to_3(self):
         with self.assertRaises(ValueError):
-            tet = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+            tet = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
             Elements(tetrahedra=tet)
 
     def tests_if_value_error_is_raised_when_tetrahedras_last_dimension_is_set_to_3(self):
-        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         with self.assertRaises(ValueError):
-            elements.tetrahedra = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+            elements.tetrahedra = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
 
     def tests_if_tetrahedra_is_none_when_initialized(self):
-        tri = torch.arange(12).reshape(4, 3)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         self.assertIsNone(elements.tetrahedra)
 
@@ -72,7 +72,7 @@ class TestElements(unittest.TestCase):
             elements = Elements(triangles=tri)
 
     def tests_if_value_error_is_raised_when_position_dtype_is_set_to_torch_float32(self):
-        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int64)
+        tri = torch.arange(12).reshape(4, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         with self.assertRaises(ValueError):
             elements.triangles = torch.arange(12).reshape(4, 3).to(dtype=torch.float32)
@@ -83,18 +83,18 @@ class TestElements(unittest.TestCase):
             Elements(tetrahedra=tet)
 
     def tests_if_value_error_is_raised_when_position_dtype_is_set_to_torch_float32(self):
-        tet = torch.arange(12).reshape(3, 4).to(dtype=torch.int64)
+        tet = torch.arange(12).reshape(3, 4).to(dtype=torch.int32)
         elements = Elements(tetrahedra=tet)
         with self.assertRaises(ValueError):
             elements.triangles = torch.arange(12).reshape(4, 3).to(dtype=torch.float32)
 
     def tests_the_num_of_triangles(self):
-        tri = torch.arange(21).reshape(7, 3).to(dtype=torch.int64)
+        tri = torch.arange(21).reshape(7, 3).to(dtype=torch.int32)
         elements = Elements(triangles=tri)
         self.assertEqual(elements.num_triangles, 7)
 
     def tests_the_num_of_tetrahedra(self):
-        tet = torch.arange(24).reshape(6, 4).to(dtype=torch.int64)
+        tet = torch.arange(24).reshape(6, 4).to(dtype=torch.int32)
         elements = Elements(tetrahedra=tet)
         self.assertEqual(elements.num_tetrahedra, 6)
